switched to using int64_t for integer representation to support bigger numbers

+++ VarCache.cpp
@@ -11,7 +11,7 @@
 namespace jasl {
 
     /// caches for ints, bools and doubles
-    std::map<std::string, int> VarCache::intCache;
+    std::map<std::string, int64_t> VarCache::intCache;
     std::map<std::string, bool> VarCache::boolCache;
     std::map<std::string, double> VarCache::doubleCache;
     std::map<std::string, std::string> VarCache::stringCache;
