+++ VarCache.hpp
@@ -11,13 +11,14 @@
 #include <map>
 #include <string>
 #include <vector>
+#include <cstdint>
 
 namespace jasl {
 
     struct VarCache 
     {
         /// caches for ints, bools and doubles
-        static std::map<std::string, int> intCache;
+        static std::map<std::string, int64_t> intCache;
         static std::map<std::string, bool> boolCache;
         static std::map<std::string, double> doubleCache;
         static std::map<std::string, std::string> stringCache;
